requirements.txt
Django==4.2.11
asgiref==3.7.2
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-tenants==3.6.1
Pillow==8.2.0
    # via 
    # django-filer
psycopg2==2.9.9
python-dateutil==2.8.1
pytz==2020.1
six==1.16.0
sqlparse==0.4.1
urllib3==2.1.0
django_compressor==4.4
# django-filer==3.1.1
# easy-thumbnails[svg]==2.8.5
django-htmlmin==0.11.0
    # via 
    # django-filer
django-polymorphic==3.1.0
    # via 
    # django-filer
# django-ckeditor-filebrowser-filer==0.5.0
    # via 
    # django-ckedtior
dnspython==2.4.2
pycryptodome==3.20.0

django_compressor==4.4
django-libsass
requests
openpyxl
unidecode
psycopg2-binary
# django-redis-cache
django-redis
# django-cors-headers
bleach